@dgarske 2024-05-22
@SparkiDev I've got a few RISC-V targets here now, so I will try this on actual HW.