@dgarske 2024-05-29
> ./configure --enable-all --enable-riscv-asm
> 
> ```
> wolfcrypt/src/aes.c: In function '_AesXtsHelper':
> wolfcrypt/src/aes.c:12631:16: error: implicit declaration of function '_AesEcbEncrypt'; did you mean 'wc_AesEcbEncrypt'? [-Werror=implicit-function-declaration]
> 12631 |         return _AesEcbEncrypt(aes, out, out, totalSz);
>       |                ^~~~~~~~~~~~~~
>       |                wc_AesEcbEncrypt
> wolfcrypt/src/aes.c:12631:16: error: nested extern declaration of '_AesEcbEncrypt' [-Werror=nested-externs]
> wolfcrypt/src/aes.c:12634:16: error: implicit declaration of function '_AesEcbDecrypt'; did you mean 'wc_AesEcbDecrypt'? [-Werror=implicit-function-declaration]
> 12634 |         return _AesEcbDecrypt(aes, out, out, totalSz);
>       |                ^~~~~~~~~~~~~~
>       |                wc_AesEcbDecrypt
> wolfcrypt/src/aes.c:12634:16: error: nested extern declaration of '_AesEcbDecrypt' [-Werror=nested-externs]
> ```

@SparkiDev  says AES XTS is not yet support with RISC-V ASM. Note: I tried to use `./configure --enable-all --disable-aesxtx --enable-riscv-asm` but that didn't work. We normally support a way to disable a specific option with all. Sean please review.