@SparkiDev 2024-06-06
Fixed --enable-all to work.