@gojimmypi 2024-06-14
@dgarske - question on the benchmarks fix data size vs fixed time:

In  the `master` and `riscv_aes_asm` branch you [ran](https://github.com/wolfSSL/wolfssl/pull/7569#issuecomment-2132248155) these commands, respectively:

```
# before, on master
./configure —enable-all
```
vs

```
# after, with ASM Optimization
./configure --enable-all --enable-riscv-asm
```

Then for comparison your ran this for both:
```
./wolfcrypt/benchmark/benchmark -aes-cbc -aes-gcm
```

The output on `master` took a fixed 5MB chunk of data and timed the completion: in this example 12.798 seconds:

```
AES-128-CBC-enc              5 MiB took 12.798 seconds,    0.391 MiB/s
```

The output on  `riscv_aes_asm` completed as soon as reasonable after a fixed one second duration and determined the amount of data processed:

```
AES-128-CBC-enc             20 MiB took 1.076 seconds,   18.588 MiB/s
```

Why the difference in fixed data size vs fixed time? 

Additionally, perhaps just nit-picky, but curious: it appears there was also difference in `bits` size generated. `bits=4096`:

```
# master
Math:   Multi-Precision: Wolf(SP) word-size=64 bits=4096 sp_int.c
```
vs `bits=3072`

```
# ASM
Math:   Multi-Precision: Wolf(SP) word-size=64 bits=3072 sp_int.c
```


It appears that `./configure --enable-all --enable-riscv-asm` produces different `user_settings.h`  than `./configure —enable-all` affecting more than just assembly optimization.  Perhaps it should be consistent, at least for the benchmark configuration? I'm also left wondering for a real apples-to-apples if master was set to `bits=3072` whether there would be a performance difference?

In any case - that's an astonishing performance boost by @SparkiDev :)