@SparkiDev 2024-06-17
The size of data processed is the number of 1048576 byte (=1MB) buffers encrypted/decrypted.
We do a minimum number of buffers regardless of platform but no less than for 1 second.

I have no idea why the number of bits in SP changed though.